refactor: move onLine to test helpers
